+++ Cesium3DTilesContent/include/Cesium3DTilesContent/GltfConverterResult.h
@@ -5,7 +5,12 @@
 #include <CesiumGltf/Model.h>
 #include <CesiumUtility/ErrorList.h>
 
+#include <glm/common.hpp>
+#include <glm/gtx/quaternion.hpp>
+
 #include <optional>
+#include <string>
+#include <vector>
 
 namespace Cesium3DTilesContent {
 /**