+++ Cesium3DTilesContent/include/Cesium3DTilesContent/I3dmToGltfConverter.h
@@ -1,11 +1,11 @@
 #pragma once
 
-#include "GltfConverterResult.h"
-
+#include <Cesium3DTilesContent/GltfConverterResult.h>
 #include <CesiumAsync/Future.h>
 #include <CesiumGltf/Model.h>
 #include <CesiumGltfReader/GltfReader.h>
 
+#include <glm/mat4x4.hpp>
 #include <gsl/span>
 
 #include <optional>
@@ -13,9 +13,9 @@
 namespace Cesium3DTilesContent {
 struct AssetFetcher;
 
-struct PntsToGltfConverter {
+struct I3dmToGltfConverter {
   static CesiumAsync::Future<GltfConverterResult> convert(
-      const gsl::span<const std::byte>& pntsBinary,
+      const gsl::span<const std::byte>& instancesBinary,
       const CesiumGltfReader::GltfReaderOptions& options,
       const AssetFetcher& assetFetcher);
 };